Replace ContainerSpec.Behavior with ContainerTests.Behavior.
